Fixed include order, a bunch of GCC 3.3 warnings, OS/2 build.
git-svn-id: http://www.virtualbox.org/svn/vbox/trunk@13908 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Runtime/generic/RTMpCpuIdToSetIndex-generic.cpp
@@ -37,6 +37,6 @@
 
 RTDECL(int) RTMpCpuIdToSetIndex(RTCPUID idCpu)
 {
-    return idCpu != NIL_RTCPUID ? idCpu : -1;
+    return idCpu != NIL_RTCPUID ? (int) idCpu : -1;
 }
